refactor(warnings): narrowing conversion

+++ ImageLounge/src/DkCore/DkMetaData.cpp
@@ -778,7 +778,7 @@ QImage DkMetaDataT::getPreviewImage(int minPreviewWidth) const
         for (size_t idx = 0; idx < pList.size(); idx++) {
             if (pList[idx].width_ > (uint32_t)maxWidth && pList[idx].width_ > (uint32_t)minPreviewWidth) {
                 mIdx = (int)idx;
-                maxWidth = pList[idx].width_;
+                maxWidth = (int)pList[idx].width_;
             }
         }
